codeclimate reducing module size
