+++ services/director-v2/src/simcore_service_director_v2/models/schemas/dynamic_services/__init__.py
@@ -1,0 +1,2 @@
+from .monitor import *
+from .service import *
